% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bCounter.R
\name{bCounter}
\alias{bCounter}
\title{A function that counts bouts within a vector}
\usage{
bCounter(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector with counted bouts
}
\description{
A function that counts bouts within a vector
}
\examples{
x = c(rep(1, 5), rep(0, 2), rep(1, 4), rep(0, 2))
bCounter(x)

data.frame(x = x,
           bout = bCounter(x))
}
