% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemp.R
\name{getTemp}
\alias{getTemp}
\title{get Temperature (data)}
\usage{
getTemp(x, y, t, PROJ)
}
\arguments{
\item{x}{Longitude of the point}

\item{y}{Latitude of the point}

\item{t}{Name of the variable}

\item{PROJ}{Projection of the point and enviromental data (has to be the same)}
}
\value{
A list containing the temperature for this point
}
\description{
This function connects points with the closest temperature data in space.
}
