% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWind.R
\name{getWind}
\alias{getWind}
\title{get Wind (data)}
\usage{
getWind(x, y, w, PROJ)
}
\arguments{
\item{x}{Longitude of the point}

\item{y}{Latitude of the point}

\item{w}{Data table with wind data containing u & v wind component and the datetime}

\item{PROJ}{Projection of the point and wind data (has to be the same)}
}
\value{
A list containing the u and v component for the point
}
\description{
This function connects points with the closest wind data in space and time.
}
