% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESA_example_tracks.R
\docType{data}
\name{PESA_example_tracks}
\alias{PESA_example_tracks}
\title{Example Pectoral Sandpiper tracks}
\format{A data.table with some short example tracks around Barrow
\itemize{
  \item bird_patch_id: Tag ID and number of left tenure location with this track
  \item datetime_: Date and time of the position
  \item x: Latitude
  \item y: Longitude
  \item year_: Year
  }}
\usage{
data(PESA_example_tracks)
}
\description{
Write a description of the data set
}
\examples{
PESA_example_tracks

}
\references{
Still to come
}
\keyword{datasets}
