% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeAlong.R
\name{sizeAlong}
\alias{sizeAlong}
\title{Creates different size values along a vector}
\usage{
sizeAlong(x, head = 20, to = c(0.1, 2.5))
}
\arguments{
\item{x}{Vector along which alpha is created}

\item{head}{Numeric parameter influencing the lenght of the head}

\item{to}{Numeric vector including the minimum and maximum size}
}
\value{
Numeric verctor with different size values
}
\description{
Creates different size values along a vector
}
\examples{
library(ggplot2)
d = data.frame(x = 1:100, y = 1:100, s = sizeAlong(1:100, head = 70, to = c(0.1, 5)))
bm =  ggplot(d, aes(x, y))
bm + geom_path(size = 1)
bm + geom_path(size = d$s, lineend = 'round')
}
