\name{w.start}
\alias{w.start}
\title{
Start Wind Data API, login to Wind Server. 
}
\description{
   To Start Wind Data API.\cr
   w.start(waitTime,showmenu)\cr

   Description: \cr
      waitTime     the maximum duration (s, default 300s) for waiting API response.\cr
      showmenu     whether to show the guide dialog(default to show it). \cr
}
\usage{
w.start(waitTime = 300, showmenu = TRUE)
}

\keyword{ start }
